fix: email matching failure during canvas grade syncs (#590)
* fix: email matching failure during canvas grade syncs

In `get_student_id_by_email`, the Canvas search API does a
case-insensitive search and returns the correct user object.
However, the function failed to return the Canvas ID of the
student as the emails capitalized in Canvas.

This commit converts both emails to lowercase during the
comparison to overcome this mismatch and returns the correct
user id.

* docs: update version and changelog